Made the advised changes for adding and removing a user from a group in group/routes
